Change code block font size to fit line numbers
The original PaperMod theme is designed without showing the line numbers
in code blocks, which is probably because of its idea of simplicity. In
order to free out line numbers one could not only enable it but also
need to adjust the CSS for not pushing the code body to the right,
avoiding possible masking of content.

- The width of the body of the post is and will continue to be 720px.
- The actual width of a single character of the font family
  SauceCodePro NFM under the current font size is 8.82px.
- The width of margin on either side is and will continue to be 7.2px.
- The width of the line number column excluding the number itself
  (which is of the same font family of the code) is and will continue to
  be 12px (padding) + 5.88px (margin).
- The width of a character to which the font should be changed is
  8.28578px = (720 - 12 - 5.88 - 7.2 * 2) / 83.
- The ratio between the width of characters and the font size is 3 : 5.
- The font size to which the font should be changed is therefore
  13.8px = 8.28578px * 5 / 3.